Moved all factories to the specific package
